paths: Move mtime test utils out of module
